Moved NotificationManager to common and created notification tasks
- Seperated Notifications to NotificationTasks
- Seperate ZMS and common functionality in Notifications
- Moved NotificationManager to Common
- Created DomainRoleMembersFetcher interface to be used for getting role members
from different sources (ZMS DB / ZTS cache)
